Added documentation for definition API. Converted int Ids to long in DB.

+++ backend/src/main/java/com/ideahub/model/Idea.java
@@ -46,12 +46,12 @@ public class Idea {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private int id;
+    private long id;
     
     // I think we need to reference just the id here instead of the object
     // as we'd run into a double-binding issue in a previous project
     @Column(name = "user_id", nullable = false)
-    private Long userId;
+    private long userId;
         
     @OneToMany(cascade = { CascadeType.ALL })
     @JoinColumn(name = "idea_id") // Which column in the referenced table will be joined
